[Loom] Separate lowering rule interpreters
Split the generated lowering-rule schema, source-value projection, match
interpreter, and emit interpreter into explicit components with
documented contracts and focused tests.

The matcher owns its descriptor-resolution cache while the shared
lowering context retains only opaque state. Planning, reporting, and
target adapters now include only the interpreter contracts they consume,
preserving immutable generated tables without a catch-all implementation
boundary.
