+++ loom/src/loom/codegen/low/lower/lower.c
@@ -14,8 +14,8 @@
 #include "loom/codegen/low/lower/function_boundary.h"
 #include "loom/codegen/low/lower/lower_context.h"
 #include "loom/codegen/low/lower/lower_report.h"
+#include "loom/codegen/low/lower/lower_rule_emit.h"
 #include "loom/codegen/low/lower/lower_rule_source_memory.h"
-#include "loom/codegen/low/lower/lower_rules.h"
 #include "loom/codegen/low/lower/source_plan.h"
 #include "loom/codegen/low/lower/source_query.h"
 #include "loom/codegen/low/source_memory_plan.h"
