+++ loom/src/loom/codegen/low/lower/lower_context.h
@@ -35,14 +35,8 @@ extern "C" {
 // Sentinel indicating that a source value was intentionally erased.
 #define LOOM_LOW_LOWER_VALUE_ID_ELIDED ((loom_value_id_t)(UINT32_MAX - 1))
 
-typedef struct loom_low_lower_rule_descriptor_map_t {
-  // Rule set whose local descriptor refs are resolved by descriptors.
-  const loom_low_lower_rule_set_t* rule_set;
-  // Descriptor rows indexed by rule-set-local descriptor ref.
-  const loom_low_descriptor_t* const* descriptors;
-  // Number of entries in descriptors.
-  uint16_t descriptor_count;
-} loom_low_lower_rule_descriptor_map_t;
+typedef struct loom_low_lower_rule_descriptor_map_t
+    loom_low_lower_rule_descriptor_map_t;
 
 typedef struct loom_low_lower_successor_interpositions_t {
   // Effective low destinations indexed by source terminator successor ordinal.
@@ -165,13 +159,6 @@ struct loom_low_lower_context_t {
 // Returns the source function name used in source-to-low diagnostics/reports.
 iree_string_view_t loom_low_lower_context_function_name(
     const loom_low_lower_context_t* context);
-
-// Resolves descriptor refs through the lowering context's cached rule maps.
-iree_status_t loom_low_lower_rule_match_descriptor_ref_from_lowering(
-    void* user_data, const loom_low_lower_rule_match_context_t* match_context,
-    const loom_low_lower_rule_set_t* rule_set,
-    loom_low_lower_descriptor_ref_t descriptor_ref,
-    const loom_low_descriptor_t** out_descriptor);
 
 // Returns true when the lowering context has reached its diagnostic limit.
 bool loom_low_lower_context_should_stop(
