+++ loom/src/loom/codegen/low/lower/lower_rule_source_memory_match.c
@@ -5,6 +5,7 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "iree/base/internal/math.h"
+#include "loom/codegen/low/lower/lower_rule_match.h"
 #include "loom/codegen/low/lower/lower_rule_source_memory.h"
 #include "loom/ir/module.h"
 #include "loom/ops/buffer/ops.h"
