+++ loom/src/loom/codegen/low/lower/lower_rule_source_memory_materialize.c
@@ -7,6 +7,7 @@
 #include "iree/base/internal/math.h"
 #include "loom/codegen/low/lower/lower_context.h"
 #include "loom/codegen/low/lower/lower_rule_descriptor.h"
+#include "loom/codegen/low/lower/lower_rule_match.h"
 #include "loom/codegen/low/lower/lower_rule_source_memory.h"
 #include "loom/ir/module.h"
 #include "loom/ops/low/ops.h"
