+++ loom/src/loom/codegen/low/lower/lower_rules.h
@@ -4,13 +4,11 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-// Table-driven source-to-target-low lowering rules.
+// Immutable generated source-to-target-low lowering-rule table schema.
 //
-// Targets use these tables to describe the common case: source op guards over
-// operands, results, attributes, and descriptor availability followed by one or
-// more descriptor-backed low packet emissions. The interpreter owns the shared
-// mechanics so target packages can grow generated .rodata instead of per-op
-// callback dispatchers.
+// Targets use these records to describe source op guards and descriptor-backed
+// Low packet emission as compact .rodata. Matching and emission are separate
+// interpreters declared by lower_rule_match.h and lower_rule_emit.h.
 
 #ifndef LOOM_CODEGEN_LOW_LOWER_LOWER_RULES_H_
 #define LOOM_CODEGEN_LOW_LOWER_LOWER_RULES_H_
@@ -25,12 +23,6 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-typedef struct loom_low_lower_rule_match_context_t
-    loom_low_lower_rule_match_context_t;
-typedef struct loom_low_lower_rule_source_memory_state_t
-    loom_low_lower_rule_source_memory_state_t;
-typedef struct loom_symbolic_expr_context_t loom_symbolic_expr_context_t;
 
 typedef uint16_t loom_low_lower_descriptor_ref_t;
 
@@ -122,86 +114,6 @@ typedef struct loom_low_lower_value_materializer_t {
   // Emission-time callback that returns the low value used by descriptor ops.
   loom_low_lower_materialize_value_fn_t materialize;
 } loom_low_lower_value_materializer_t;
-
-typedef iree_status_t (*loom_low_lower_rule_match_map_value_fn_t)(
-    void* user_data, const loom_low_lower_rule_match_context_t* context,
-    const loom_op_t* source_op, loom_value_id_t source_value_id,
-    loom_low_lower_rule_mapped_value_t* out_mapped_value);
-
-typedef struct loom_low_lower_rule_match_map_value_callback_t {
-  // Callback invoked to map one source value into target-low register metadata.
-  loom_low_lower_rule_match_map_value_fn_t fn;
-  // Caller-owned payload passed to |fn|.
-  void* user_data;
-} loom_low_lower_rule_match_map_value_callback_t;
-
-typedef iree_status_t (*loom_low_lower_rule_match_can_materialize_value_fn_t)(
-    void* user_data, const loom_low_lower_rule_match_context_t* context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    uint16_t value_ref_index, loom_value_id_t source_value_id,
-    bool* out_can_materialize);
-
-typedef struct loom_low_lower_rule_match_can_materialize_value_callback_t {
-  // Callback invoked for VALUE_MATERIALIZABLE guards.
-  loom_low_lower_rule_match_can_materialize_value_fn_t fn;
-  // Caller-owned payload passed to |fn|.
-  void* user_data;
-} loom_low_lower_rule_match_can_materialize_value_callback_t;
-
-typedef iree_status_t (*loom_low_lower_rule_match_descriptor_ref_fn_t)(
-    void* user_data, const loom_low_lower_rule_match_context_t* context,
-    const loom_low_lower_rule_set_t* rule_set,
-    loom_low_lower_descriptor_ref_t descriptor_ref,
-    const loom_low_descriptor_t** out_descriptor);
-
-typedef struct loom_low_lower_rule_match_descriptor_ref_callback_t {
-  // Callback invoked to resolve a rule-set-local descriptor ref.
-  loom_low_lower_rule_match_descriptor_ref_fn_t fn;
-  // Caller-owned payload passed to |fn|.
-  void* user_data;
-} loom_low_lower_rule_match_descriptor_ref_callback_t;
-
-typedef uint16_t loom_low_lower_rule_match_flags_t;
-
-// Match contract-only rule rows that are visible to read-only legality queries
-// but never execute as source-to-low emission programs.
-#define LOOM_LOW_LOWER_RULE_MATCH_FLAG_CONTRACT_ONLY \
-  ((loom_low_lower_rule_match_flags_t)1u << 0)
-
-struct loom_low_lower_rule_match_context_t {
-  // Source module being matched.
-  const loom_module_t* module;
-  // Source function whose body is being matched.
-  loom_func_like_t function;
-  // Target bundle selected for this match.
-  const loom_target_bundle_t* bundle;
-  // Descriptor set selected for the target-low contract.
-  const loom_low_descriptor_set_t* descriptor_set;
-  // Feature bits selected by the target-low contract.
-  uint64_t feature_bits;
-  // Source-value to target-low register metadata mapper.
-  loom_low_lower_rule_match_map_value_callback_t map_value;
-  // Optional source value materializer predicate bridge.
-  loom_low_lower_rule_match_can_materialize_value_callback_t can_materialize;
-  // Optional rule-local descriptor-ref resolver. Missing uses descriptor keys
-  // directly and is intended for tests and cold standalone queries.
-  loom_low_lower_rule_match_descriptor_ref_callback_t descriptor_ref;
-  // Optional dense source value facts used by fact-backed guard rows.
-  const loom_value_fact_table_t* fact_table;
-  // Optional precomputed view summaries used by source-memory guard rows.
-  const loom_view_region_table_t* view_regions;
-  // Caller-owned state retaining the canonical source-memory plan for the
-  // source op being matched. Required for rule sets with source-memory rows.
-  loom_low_lower_rule_source_memory_state_t* source_memory_state;
-  // Optional symbolic proof context used as a cold fallback for fact-backed
-  // guard rows whose scalar intervals are inconclusive.
-  loom_symbolic_expr_context_t* symbolic_expr_context;
-  // Match behavior flags.
-  loom_low_lower_rule_match_flags_t flags;
-  // One-based policy rule-set ordinal supplied by composed contract selection;
-  // zero when no policy owner is known.
-  uint16_t policy_rule_set_ordinal;
-};
 
 typedef struct loom_low_lower_rule_descriptor_ref_t {
   // Rule-set B-string offset for the stable descriptor key.
@@ -807,13 +719,6 @@ typedef struct loom_low_lower_emit_t {
   uint16_t source_memory_ordinal;
 } loom_low_lower_emit_t;
 
-typedef struct loom_low_lower_resolved_emit_t {
-  // Static emit-program row selected by planning.
-  const loom_low_lower_emit_t* emit;
-  // Descriptor row referenced by |emit| and resolved during planning.
-  loom_low_lower_resolved_descriptor_t descriptor;
-} loom_low_lower_resolved_emit_t;
-
 typedef uint16_t loom_low_lower_rule_flags_t;
 
 // Rule row is a read-only target contract case and must not be selected as an
@@ -945,141 +850,13 @@ typedef struct loom_low_lower_rule_set_t {
   uint16_t diagnostic_count;
 } loom_low_lower_rule_set_t;
 
-// Returns the rule-set B-string view at |string_offset|. Generated rule sets
-// and offsets are trusted compiler-owned tables.
-iree_string_view_t loom_low_lower_rule_set_string(
+// Returns the trusted rule-set B-string at |string_offset|.
+static inline iree_string_view_t loom_low_lower_rule_set_string(
     const loom_low_lower_rule_set_t* rule_set,
-    loom_bstring_table_offset_t string_offset);
-
-typedef struct loom_low_lower_rule_selection_t {
-  // Selected rule row, or NULL when no rule accepted the source op.
-  const loom_low_lower_rule_t* rule;
-  // Selected rule row ordinal, or UINT16_MAX when no rule accepted the source
-  // op.
-  uint16_t rule_index;
-  // True when the rule set had at least one rule span for the source op kind.
-  bool has_source_op_span;
-  // Diagnostic row describing the best failed guard when |rule| is NULL.
-  uint16_t diagnostic_index;
-  // Number of guards matched by the best failed rule candidate.
-  uint16_t matched_guard_count;
-  // True when a descriptor guard rejected a rule whose source-memory
-  // constraints matched the source access.
-  bool source_memory_compatible;
-  // True when the selected rule consumes the canonical source-memory plan.
-  bool uses_source_memory_access;
-} loom_low_lower_rule_selection_t;
-
-// Initializes a rule match context backed by a mutable lowering context.
-// |source_memory_state| retains the canonical source-memory plan across every
-// candidate rule inspected for one source op.
-void loom_low_lower_rule_match_context_initialize_from_lowering(
-    loom_low_lower_context_t* context,
-    const loom_view_region_table_t* view_regions,
-    loom_low_lower_rule_source_memory_state_t* source_memory_state,
-    loom_low_lower_rule_match_context_t* out_match_context);
-
-// Selects the exact lowering rule for |source_op| without emitting user
-// diagnostics. Callers that compose rule tables with custom target callbacks
-// can use the recorded failure detail if every lowering path rejects the op.
-iree_status_t loom_low_lower_rule_set_select(
-    loom_low_lower_context_t* context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    loom_low_lower_rule_selection_t* out_selection);
-
-// Selects the exact target contract rule for |source_op| using the mutable
-// lowering context. Unlike loom_low_lower_rule_set_select this includes
-// contract-only rows and is intended for target recipe selectors that consume
-// generated contracts but still need target-owned emit code.
-iree_status_t loom_low_lower_rule_set_select_contract(
-    loom_low_lower_context_t* context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    loom_low_lower_rule_selection_t* out_selection);
-
-// Selects a lowering rule from a caller-selected rule range using the mutable
-// lowering context. Contract-table source lowering uses this after direct
-// op-kind lookup has selected the candidate row.
-iree_status_t loom_low_lower_rule_set_select_rule_range(
-    loom_low_lower_context_t* context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    uint16_t rule_start, uint16_t rule_count,
-    loom_low_lower_rule_selection_t* out_selection);
-
-// Selects the exact lowering rule for |source_op| using a read-only match
-// context. This is the legality/checking sibling of
-// loom_low_lower_rule_set_select and performs no IR mutation or emission.
-iree_status_t loom_low_lower_rule_set_select_with_match_context(
-    const loom_low_lower_rule_match_context_t* match_context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    loom_low_lower_rule_selection_t* out_selection);
-
-// Selects the exact lowering rule for |source_op| from a caller-selected rule
-// range using a read-only match context. Contract-table systems use this after
-// direct op-kind lookup has selected the candidate span.
-iree_status_t loom_low_lower_rule_set_select_rule_range_with_match_context(
-    const loom_low_lower_rule_match_context_t* match_context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    uint16_t rule_start, uint16_t rule_count,
-    loom_low_lower_rule_selection_t* out_selection);
-
-// Returns the diagnostic row for |selection|, or NULL when no table diagnostic
-// is available for that failed selection.
-const loom_low_lower_diagnostic_t* loom_low_lower_rule_set_selection_diagnostic(
-    const loom_low_lower_rule_set_t* rule_set,
-    loom_low_lower_rule_selection_t selection);
-
-// Materializes generated diagnostic parameter projections for a rejected rule.
-void loom_low_lower_rule_materialize_diagnostic_params(
-    const loom_low_lower_rule_match_context_t* match_context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    const loom_low_lower_diagnostic_t* diagnostic,
-    loom_diagnostic_param_t* out_params);
-
-// Resolves an operand/result value-ref row to the flat source-op value span for
-// that field. Non-source value refs return an empty span.
-loom_value_slice_t loom_low_lower_rule_value_ref_field_span(
-    const loom_module_t* module, const loom_low_lower_rule_set_t* rule_set,
-    const loom_op_t* source_op, uint16_t value_ref_index);
-
-// Resolves a rule-set-local descriptor ref against |match_context|'s selected
-// descriptor set. Missing optional descriptors return NULL.
-iree_status_t loom_low_lower_rule_resolve_descriptor_ref(
-    const loom_low_lower_rule_match_context_t* match_context,
-    const loom_low_lower_rule_set_t* rule_set,
-    loom_low_lower_descriptor_ref_t descriptor_ref,
-    const loom_low_descriptor_t** out_descriptor);
-
-// Returns the first descriptor ref emitted by |rule|, or
-// LOOM_LOW_LOWER_DESCRIPTOR_REF_NONE when the rule does not emit a
-// descriptor-backed packet.
-loom_low_lower_descriptor_ref_t loom_low_lower_rule_first_descriptor_ref(
-    const loom_low_lower_rule_set_t* rule_set,
-    const loom_low_lower_rule_t* rule);
-
-// Emits the diagnostic described by a failed selection. If the rule set did not
-// cover the source op kind, emits the generic no-mapping diagnostic.
-iree_status_t loom_low_lower_rule_set_emit_selection_failure(
-    loom_low_lower_context_t* context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    loom_low_lower_rule_selection_t selection,
-    loom_low_lower_rule_source_memory_state_t* source_memory_state);
-
-// Resolves descriptor-backed emit rows for |rule| after selection. The returned
-// rows are arena-owned by |context| and remain valid for the current lowering
-// run.
-iree_status_t loom_low_lower_rule_set_resolve_emit_program(
-    loom_low_lower_context_t* context, uint16_t rule_set_index,
-    const loom_low_lower_rule_set_t* rule_set,
-    const loom_low_lower_rule_t* rule,
-    const loom_low_lower_resolved_emit_t** out_resolved_emits);
-
-// Emits target-low packets for |source_op| using a previously selected rule.
-iree_status_t loom_low_lower_rule_set_emit_rule(
-    loom_low_lower_context_t* context,
-    const loom_low_lower_rule_set_t* rule_set, const loom_op_t* source_op,
-    const loom_low_lower_rule_t* rule,
-    const loom_low_lower_resolved_emit_t* resolved_emits,
-    const loom_low_source_memory_access_plan_t* source_memory_access);
+    loom_bstring_table_offset_t string_offset) {
+  return loom_bstring_view(
+      loom_bstring_table_get(&rule_set->string_table, string_offset));
+}
 
 #ifdef __cplusplus
 }  // extern "C"
