+++ loom/src/loom/codegen/low/lower/source_plan.c
@@ -11,7 +11,10 @@
 #include "loom/analysis/contract_vector.h"
 #include "loom/codegen/low/lower/contract_query.h"
 #include "loom/codegen/low/lower/lower_context.h"
+#include "loom/codegen/low/lower/lower_rule_emit.h"
+#include "loom/codegen/low/lower/lower_rule_match.h"
 #include "loom/codegen/low/lower/lower_rule_source_memory.h"
+#include "loom/codegen/low/lower/lower_rule_value.h"
 #include "loom/codegen/low/lower/source_query.h"
 #include "loom/codegen/low/source_memory_plan.h"
 #include "loom/error/error_catalog.h"
